Aperfeiçoando a interface de Categoria/index.

+++ FirstApplication.ASPdotNET.MVC5.Books/Models/Categoria.cs
@@ -2,7 +2,7 @@
 {
     public class Categoria
     {
-        public int CategoriaId { get; set; }
+        public long CategoriaId { get; set; }
         public string Nome { get; set; }
     }
 }
